Refactoring existing animation functions
