+++ packages/framer-motion/src/animation/hooks/use-animated-state.ts
@@ -3,10 +3,10 @@ import { useConstant } from "../../utils/use-constant"
 import { checkTargetForNewValues, getOrigin } from "../../render/utils/setters"
 import { TargetAndTransition } from "../../types"
 import { ResolvedValues } from "../../render/types"
-import { animateVisualElement } from "../../render/utils/animation"
 import { makeUseVisualState } from "../../motion/utils/use-visual-state"
 import { createBox } from "../../projection/geometry/models"
 import { VisualElement } from "../../render/VisualElement"
+import { animateVisualElement } from "../interfaces/visual-element"
 
 interface AnimatedStateOptions {
     initialState: ResolvedValues
