+++ packages/framer-motion/src/animation/interfaces/types.ts
@@ -1,0 +1,9 @@
+import { AnimationType } from "../../render/utils/types"
+import { Transition } from "../../types"
+
+export type VisualElementAnimationOptions = {
+    delay?: number
+    transitionOverride?: Transition
+    custom?: any
+    type?: AnimationType
+}
