+++ packages/framer-motion/src/animation/interfaces/visual-element-target.ts
@@ -1,0 +1,107 @@
+import { sync } from "../../frameloop"
+import { transformProps } from "../../render/html/utils/transform"
+import type { AnimationTypeState } from "../../render/utils/animation-state"
+import type { VisualElement } from "../../render/VisualElement"
+import type { TargetAndTransition } from "../../types"
+import { optimizedAppearDataAttribute } from "../optimized-appear/data-id"
+import type { VisualElementAnimationOptions } from "./types"
+import { animateMotionValue } from "./motion-value"
+import { isWillChangeMotionValue } from "../../value/use-will-change/is"
+import { setTarget } from "../../render/utils/setters"
+
+/**
+ * Decide whether we should block this animation. Previously, we achieved this
+ * just by checking whether the key was listed in protectedKeys, but this
+ * posed problems if an animation was triggered by afterChildren and protectedKeys
+ * had been set to true in the meantime.
+ */
+function shouldBlockAnimation(
+    { protectedKeys, needsAnimating }: AnimationTypeState,
+    key: string
+) {
+    const shouldBlock =
+        protectedKeys.hasOwnProperty(key) && needsAnimating[key] !== true
+
+    needsAnimating[key] = false
+    return shouldBlock
+}
+
+export function animateTarget(
+    visualElement: VisualElement,
+    definition: TargetAndTransition,
+    { delay = 0, transitionOverride, type }: VisualElementAnimationOptions = {}
+): Promise<any> {
+    let {
+        transition = visualElement.getDefaultTransition(),
+        transitionEnd,
+        ...target
+    } = visualElement.makeTargetAnimatable(definition)
+
+    const willChange = visualElement.getValue("willChange")
+
+    if (transitionOverride) transition = transitionOverride
+
+    const animations: Promise<any>[] = []
+
+    const animationTypeState =
+        type &&
+        visualElement.animationState &&
+        visualElement.animationState.getState()[type]
+
+    for (const key in target) {
+        const value = visualElement.getValue(key)
+        const valueTarget = target[key]
+
+        if (
+            !value ||
+            valueTarget === undefined ||
+            (animationTypeState &&
+                shouldBlockAnimation(animationTypeState, key))
+        ) {
+            continue
+        }
+
+        const valueTransition = { delay, elapsed: 0, ...transition }
+
+        /**
+         * If this is the first time a value is being animated, check
+         * to see if we're handling off from an existing animation.
+         */
+        if (window.HandoffAppearAnimations && !value.hasAnimated) {
+            const appearId =
+                visualElement.getProps()[optimizedAppearDataAttribute]
+
+            if (appearId) {
+                valueTransition.elapsed = window.HandoffAppearAnimations(
+                    appearId,
+                    key,
+                    value,
+                    sync
+                )
+            }
+        }
+
+        let animation = value.start(
+            animateMotionValue(
+                key,
+                value,
+                valueTarget,
+                visualElement.shouldReduceMotion && transformProps.has(key)
+                    ? { type: false }
+                    : valueTransition
+            )
+        )
+
+        if (isWillChangeMotionValue(willChange)) {
+            willChange.add(key)
+
+            animation = animation.then(() => willChange.remove(key))
+        }
+
+        animations.push(animation)
+    }
+
+    return Promise.all(animations).then(() => {
+        transitionEnd && setTarget(visualElement, transitionEnd)
+    })
+}
