+++ packages/framer-motion/src/render/dom/utils/parse-dom-variant.ts
@@ -1,6 +1,17 @@
+import { Target, TargetWithKeyframes } from "../../../types"
+import type { VisualElement } from "../../VisualElement"
 import { resolveCSSVariables } from "./css-variables-conversion"
-import { MakeTargetAnimatable } from "../../utils/animation"
 import { unitConversion } from "./unit-conversion"
+
+export type MakeTargetAnimatable<T = unknown> = (
+    visualElement: VisualElement<T>,
+    target: TargetWithKeyframes,
+    origin?: Target,
+    transitionEnd?: Target
+) => {
+    target: TargetWithKeyframes
+    transitionEnd?: Target
+}
 
 /**
  * Parse a DOM variant to make it animatable. This involves resolving CSS variables
