+++ packages/framer-motion/src/render/types.ts
@@ -1,11 +1,11 @@
 import { VisualState } from "../motion/utils/use-visual-state"
 import { MotionValue } from "../value"
 import { ReducedMotionConfig } from "../context/MotionConfigContext"
-import { AnimationDefinition } from "./utils/animation"
 import { Axis, Box } from "../projection/geometry/types"
 import type { VisualElement } from "./VisualElement"
 import { PresenceContextProps } from "../context/PresenceContext"
 import { MotionProps } from "../motion/types"
+import { AnimationDefinition } from "../animation/types"
 
 export type GenericValues = {
     [key: string]: string | number
